* removed inline in a lot of functions that shouldnt have been inlined * removed misbehaving min/max defines, use std::min/std::max! * fixed float2int32 under x64, provided optimized assembly version for vc++ x86
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2693 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/shared/NGLog.h
@@ -90,7 +90,7 @@ public:
 		};
 		fputs(colorstrings[color], stdout);
 #else
-        SetConsoleTextAttribute(stdout_handle, color);
+        SetConsoleTextAttribute(stdout_handle, (WORD)color);
 #endif
 	}
